Interpreter: Removing some types used in rapid prototyping
In order to get some foundation up and running some types like
unsigned char where used for example for memory representation.
As this is technically correct, there are better matches like
uint8_t or uint16_t that are also more expressive.

+++ Interpreter/DisplayBuffer.cpp
@@ -24,6 +24,7 @@
 #include "DisplayBuffer.h"
 #include <Types.h>
 #include <cstddef>
+#include <cstdint>
 #include <iostream>
 
 void Chip8::DisplayBuffer::apply_display_data(const unsigned short* new_display_data)
@@ -67,7 +68,7 @@ int Chip8::DisplayBuffer::get_height()
     return DISPLAY_HEIGHT;
 }
 
-unsigned short* Chip8::DisplayBuffer::get_display_data()
+uint32_t* Chip8::DisplayBuffer::get_display_data()
 {
     return m_display_data;
 }
